diplom completed

+++ ManagementDogovorami/Parts.cs
@@ -15,7 +15,7 @@ namespace ManagementDogovorami
     public partial class Parts
     {
         public int ID { get; set; }
-        public Nullable<decimal> Price { get; set; }
+        public Nullable<int> Price { get; set; }
         public Nullable<System.DateTime> Pay_day { get; set; }
         public Nullable<int> Contract_id { get; set; }
